refactor(lexer): split `classify` pipeline stage into multiple files
